requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
django-embed-video
django-filter==22.1
idna==3.3
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
waitress==2.1.2
watchdog==2.1.9
